Moving data_directory fixture so it can be reused by backend tests
